pyproject.toml
[project]
name = "dataset-converter"
version = "0.1.0"
description = "Temporal-based dataset conversion pipeline for SFT formatting"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "temporalio>=1.6.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "pydantic>=2.7.0",
    "langchain>=0.2.0",
    "langchain-nvidia-ai-endpoints>=0.1.0",
    "datasets>=4.8.4",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "ipython",
    "black",
    "ruff",
]
